+++ MOVEit.MIT/Private/Confirm-MITToken.ps1
@@ -26,11 +26,16 @@ function Confirm-MITToken {
         $elapsed = New-TimeSpan -Start $script:Token.CreatedAt
         Write-Verbose "MIT Token at $($elapsed.TotalSeconds.ToString('F0')) of $($script:Token.ExpiresIn) seconds"
 
+        # Check if the refresh token is already expired and, if so, exit with an error
+        if ($elapsed.TotalSeconds -ge $script:Token.ExpiresIn) {
+            Write-Error "MIT Token expired.  Call Connect-MITServer to reconnect." -ErrorAction Stop
+        }
+
         # If the key is half-way to expiring, let's go ahead and
         # refresh it.
         if ($elapsed.TotalSeconds -gt ($script:Token.ExpiresIn / 2)) {
 
-            Write-Verbose "MIT Token expired, refreshing..."
+            Write-Verbose "MIT Token over half expired, refreshing..."
 
             $params = @{
                 Uri = "$script:BaseUri/token"
@@ -39,6 +44,12 @@ function Confirm-MITToken {
                 Body = "grant_type=refresh_token&refresh_token=$($script:Token.RefreshToken)"
                 Headers = @{Accept = "application/json"}
                 UserAgent = $script:UserAgent
+            }
+
+            # Add SkipCertificateCheck parameter if set
+            if ($script:SkipCertificateCheck) {
+                $params['SkipCertificateCheck'] = $true
+                Write-Verbose "SkipCertificateCheck: $true"
             }
             
             $response = Invoke-RestMethod @params